feat: Add image field to std-club model
This commit adds the `img` field to the `std-club` model in order to store the URL of an image associated with each club. The field is of type `String` and is not required, with a default value of an empty string. This change allows for the storage and retrieval of club images in the application.