feat: Add stdId and honorific fields to std-club model
This commit adds the `stdId` and `honorific` fields to the `std-club` model. The `stdId` field is of type `String` and is required, while the `honorific` field is also of type `String` and is required. These fields allow for the storage of student IDs and honorifics associated with each club member, providing additional information for club management.